requirements.txt
streamlit
google-genai
pymupdf
python-dotenv
pillow
